requirements_pi.txt
opencv-python-headless==4.13.0.92
numpy==2.4.4
scipy==1.17.1
scikit-image==0.26.0
imageio==2.37.3
tifffile==2026.3.3
pillow==12.2.0
scikit-learn==1.8.0
joblib==1.5.3
onnxruntime==1.25.0
protobuf==7.34.1
flatbuffers==25.12.19
threadpoolctl==3.6.0
networkx==3.6.1
lazy-loader==0.5
matplotlib==3.10.9
weasyprint==68.1
Jinja2==3.1.6
MarkupSafe==3.0.3
fastapi==0.115.12
uvicorn==0.34.3
websockets==15.0.1
python-multipart==0.0.20
pytesseract==0.3.13
pyzbar==0.1.9
skl2onnx==1.17.0
onnx==1.17.0
psutil==7.0.0
pyyaml==6.0.2
bcrypt==4.2.0
